feat: add local data storage
